State: Enable glColorMask
The color mask state is odd because it sets state for
colors (red/blue/green/alpha) which is not a float value.  To work
around this mismatch of our data model, the values are split into
different state settings in the UI that resolve to a single GL call.

This is the first state setting that is in a nested directory.