Retrace: Filter out sync and query calls during retrace
A program may conditionally render based on the results of a query,
but a recorded trace will not make use of query results.  There is no
reason to invoke them.  When a frame is retraced from the UI, it will
constantly retrieve query objects that have already been used.

Similarly, sync objects are used in titles to ensure rendering in
previous frames are complete.  Repeating a frame will cause the render
loop to wait on sync objects that have already been used.