framemetrics: discard useless context switches
Each context switch requires a start/stop of metrics collection.  Some
titles (eg CSGO) may set the context to NULL and then switch back to
the original context, which has no affect on rendering but complicates
metrics collection.

Framemetrics should only retrace context switches that are necessary
to the rendering of the trace.