framemetrics: drain errors from trace
framemetrics asserts in a number of places that there are no gl errors,
which is good.  But sometimes the trace being replayed triggers it's own
errors, which is bad.

So before replaying a call, check that there are no errors, and after
the call drain any errors the call itself triggered.  This should ensure
that the asserts only catch errors from framemetrics itself.