framemetrics: add support to specify counters within group
With GL_AMD_performance_monitor, hw may not support collecting all
counters within a group at the same time.  Ie. there may be more
possible options of things to count than physical counters.  So add
a way to specify specific countable to collect, ie:

 -g group:counter1,counter2,counter3

Encapsulate the specified group and optional list of counters (the
default behavior is still to collect them all) in PerfMetricDescriptor
since the next step is to allow multiple groups to be collected at
the same time.