framemetrics: split out metrics
To prepare to refactor things to handle a single query with metrics from
multiple groups (ie. GL_AMD_performance_monitor), split things out.

This could also make it possible to re-use the metrics gathering in
other tools.

Also, decouple dumping of the available metrics names from
PerfMetricGroup to simplify things.  This way PerfMetricsGroup, and
eventually a PerfMetrics class which encapsulates multiple groups which
may (or may not for intel counters) be part of the same query, do not
have to deal with the case that a PerfMetricDescriptor is not specified.