framemetrics: refactor out PerfMetrics umbrella class
Refactor out a class which can contain multiple groups of metrics.
Current implementation on both the intel and amd side is a trivial
implementation supporting just one group.  This is prep work for
following patches that will enhance amd metrics to support multiple
groups of counters in a single query.